Corrections and new seeds
